+++ packages/@sanity/pkg-utils/src/node/core/config/index.ts
@@ -1,4 +1,0 @@
-export {defineConfig} from './defineConfig.ts'
-export {loadConfig} from './loadConfig.ts'
-export * from './resolveConfigProperty.ts'
-export * from './types.ts'